+++ sg3_Inputs.js
@@ -1,0 +1,14 @@
+function saveToStore() {
+    savedColorSquares.at(colorStoresSelected).colorHeld = currColor;
+    colorStores[colorStoresSelected] = currColor;
+
+}
+
+function previewScale() {
+    prevCellSize = previewSelect.value;
+}
+
+function colorText() {
+    currColor = colorTextElement.value;
+    colorPicker.color.hexString = currColor;
+}